+++ openyich-framework/src/main/java/com/openyich/framework/boot/config/JacksonConfiguration.java
@@ -2,6 +2,8 @@ package com.openyich.framework.boot.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.zalando.problem.ProblemModule;
+import org.zalando.problem.validation.ConstraintViolationProblemModule;
 
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -25,4 +27,14 @@ public class JacksonConfiguration {
     return new Jdk8Module();
   }
 
+  @Bean
+  public ProblemModule problemModule() {
+      return new ProblemModule();
+  }
+
+  @Bean
+  public ConstraintViolationProblemModule constraintViolationProblemModule() {
+      return new ConstraintViolationProblemModule();
+  }
+  
 }