+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 2.1.1
 
 - Upgrading Spring Boot Admin dependency version to 2.1.0
+- Add domain abstractAuditingEntity
+- Add org.zalando.problem-spring-web
 
 ### 2.1.0